requirements.txt
Flask==2.3.2
jwcrypto==1.4.2
keycloak_client==0.15.4
PyLD==2.0.3
Requests==2.31.0
